Avoid warning with signed/unsigned integer

+++ 05-mesh/05-mesh.cpp
@@ -109,7 +109,7 @@ public:
                     magma::clear::depthOne
                 });
             {
-                cmdBuffer->setViewport(0, 0, width, negateViewport ? -height : height);
+                cmdBuffer->setViewport(0, 0, width, negateViewport ? -int32_t(height) : height);
                 cmdBuffer->setScissor(0, 0, width, height);
                 cmdBuffer->bindDescriptorSet(wireframePipeline, 0, descriptorSet);
                 cmdBuffer->bindPipeline(wireframePipeline);
